refactor: extract _ensure_registry_for_devices to RegistryOperations
- Extract _ensure_registry_for_devices (~673 lines) to registry.py
- Add required imports: Collection, SUBENTRY_TYPE_HUB, SUBENTRY_TYPE_TRACKER
- Add helper function imports: is_hub_device_check, normalize_device_name, resolve_tracker_subentry_candidate
- main.py reduced from 6,832 to 6,159 lines
- registry.py increased from 1,256 to 1,935 lines
- Total coordinator reduction: 8,342 → 6,159 (26%)
